+++ .dockerignore
@@ -32,3 +32,4 @@ __pycache__/
 *.sql
 data/
 pgadmin/
+.local/
