+++ .gitignore
@@ -110,8 +110,6 @@ ENV/
 env.bak/
 venv.bak/
 .idea
-data/
-pgadmin/
 
 # Spyder project settings
 .spyderproject
@@ -130,10 +128,3 @@ dmypy.json
 
 # Pyre type checker
 .pyre/
-
-# Idea
-/client/cache/
-/server/cafe/media/
-/book/media/
-data
-pgadmin
