+++ authentication/accounts/admin.py
@@ -1,41 +1,11 @@
-from django import forms
 from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin as BaseUserAdmin
-from django.contrib.auth.forms import ReadOnlyPasswordHashField
-from django.contrib.auth.password_validation import validate_password
 
+from .forms import UserChangeForm, UserCreationForm
 from .models import User
 
 
-class UserCreationForm(forms.ModelForm):
-    password = forms.CharField(label='Password', widget=forms.PasswordInput)
-
-    class Meta:
-        model = User
-        fields = ('email', 'first_name')
-
-    def clean_password(self):
-        password = self.cleaned_data.get("password")
-        validate_password(password)
-        return password
-
-    def save(self, commit=True):
-        # Save the provided password in hashed format
-        user = super().save(commit=False)
-        user.set_password(self.cleaned_data["password"])
-        if commit:
-            user.save()
-        return user
-
-
-class UserChangeForm(forms.ModelForm):
-    password = ReadOnlyPasswordHashField()
-
-    class Meta:
-        model = User
-        fields = '__all__'
-
-
+@admin.register(User)
 class UserAdmin(BaseUserAdmin):
     form = UserChangeForm
     add_form = UserCreationForm
@@ -50,6 +20,3 @@ class UserAdmin(BaseUserAdmin):
         }),
     )
     filter_horizontal = ()
-
-
-admin.site.register(User, UserAdmin)
