+++ authentication/authentication/urls.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from django.urls import path, include
+from django.urls import include, path
 
 urlpatterns = [
     path('', include('accounts.urls')),
