+++ docker-compose.yml
@@ -2,8 +2,8 @@ version: '3.9'
 services:
 
   db:
-    container_name: db
-    image: postgres
+    container_name: iteen.crm.db
+    image: postgres:15.2
     networks: 
       - iteen
     restart: always
@@ -13,25 +13,25 @@ services:
       POSTGRES_USER: admin
       POSTGRES_PASSWORD: admin
       POSTGRES_DB: testbase
-      POSTGRES_INITDB_ARGS: "-A md5"  
+      POSTGRES_HOST_AUTH_METHOD: scram-sha-256
     volumes:
-      - ./data:/var/lib/postgresql/data
+      - .local/data:/var/lib/postgresql/data
     healthcheck:
       test: ["CMD-SHELL", "pg_isready -U admin -d testbase"]
-      interval: 10s
+      interval: 5s
       timeout: 5s
       retries: 5
-      start_period: 10s
+      start_period: 5s
 
   pgadmin:
-    container_name: pgadmin_container
-    image: dpage/pgadmin4
+    container_name: iteen.pgadmin
+    image: dpage/pgadmin4:7.1
     environment:
       PGADMIN_DEFAULT_EMAIL: "admin@admin.com"
       PGADMIN_DEFAULT_PASSWORD: "admin"
       PGADMIN_CONFIG_SERVER_MODE: "False"
     volumes:
-      - ./pgadmin:/var/lib/pgadmin
+      - .local/pgadmin:/var/lib/pgadmin
     ports:
       - "5050:80"
     restart: always
@@ -48,17 +48,18 @@ services:
         - PYTHON_VERSION=${WEBAPP_PYTHON_VERSION:-3.11}
         - VERSION=${WEBAPP_VERSION:-dev}
       context: .
-    container_name: webapp
+    container_name: iteen.crm.webapp
     image: webapp:latest
     networks:
       - iteen
     ports:
       - "8000:8000"
     volumes:
-      - ./:/app/:ro
+      - .local/:/app/:ro
     init: true
     depends_on:
-      - db
+      db:
+        condition: service_healthy
       
 networks:
   iteen:
